requirements.txt
# Core dependencies
dart-vm-client==0.0.1.dev4
pydantic>=2.0.0
python-dotenv>=0.21.0
react-native-debugger-client==0.0.1.dev1

# LangChain and LLM integration
langchain-core>=0.1.0
langchain-openai>=0.0.1
openai>=1.0.0
